code refactoring [MDS-771]
