+++ docs/app/components/server/checkbox/page.tsx
@@ -1,5 +1,5 @@
 import { Default } from '@/app/components/server/checkbox/examples/Default';
-import ExampleSection from '@/app/components/shared/ExampleSection';
+import ExampleSection from '@/app/components/shared/exampleSection/ExampleSection';
 import QuickNav from '@/app/components/shared/QuickNav';
 import { getExamples } from '@/app/utils/getExamples';
 import { MDX } from '@/components/MDX';
