+++ docs/app/components/shared/exampleSection/codePreview/CodePreview.tsx
@@ -1,4 +1,4 @@
-import CodePreviewWrapper from './CodePreviewWrapper';
+import CodePreviewWrapper from './wrapper/CodePreviewWrapper';
 
 const CodePreview = ({ code }: { code: string }) => {
   return (
